Make the traffic-manager service port actually configurable using Helm
It is picked being picked up from the `SERVICE_PORT` environment and
defaults to 8081, but wasn't configurable. Now it is.

+++ pkg/agentconfig/sidecar.go
@@ -125,7 +125,7 @@ type Sidecar struct {
 	ManagerHost string `json:"managerHost,omitempty"`
 
 	// The port used when connecting to the traffic manager
-	ManagerPort int32 `json:"managerPort,omitempty"`
+	ManagerPort uint16 `json:"managerPort,omitempty"`
 
 	// The port used by the agents restFUL API server
 	APIPort uint16 `json:"apiPort,omitempty"`
